# typescript/src/swagger/client-server.openapi.json
{
    "components": {
        "schemas": {
...
            "PlaidGetKycReq": {
                "properties": {
                    "fields": {
                        "items": {
                            "$ref": "#/components/schemas/PlaidIdvField"
                        },
                        "type": "array"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "user_id"
                ],
                "type": "object"
            },
            "PlaidGetResultResp": {
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/PlaidGetResultResp"
                    },
                    {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object"
                    }
                ]
            },
...
            "PlaidStartIdvRequest": {
                "properties": {
                    "callback_url": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "user_id",
                    "email",
                    "callback_url"
                ],
                "type": "object"
            },
            "PlaidStartIdvResp": {
                "properties": {
                    "start_idv_uri": {
                        "type": "string"
                    }
                },
                "required": [
                    "start_idv_uri"
                ],
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "idv-client-server",
        "version": "0.1.0.0"
    },
    "openapi": "3.0.0",
    "paths": {
        "/v1/idv/ca/cookie/start": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaidStartIdvRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaidStartIdvResp"
                                }
                            }
                        },
                        "description": "",
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid `body`"
                    }
                }
            }
        },
...
        "/v1/idv/ca/kyc/get": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaidGetKycReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaidGetResultResp"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/ca/kyc/put": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaidPutKycReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {}
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body`"
                    }
                }
            }
        },
        "/v1/idv/ca/start": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaidStartIdvRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaidStartIdvResp"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
...
        "/v1/idv/cn/kyc/get": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvGetResultReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonValue"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/cn/mock/kyc/get": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvMockGetResultReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonValue"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/cn/mock/start": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvMockStartReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/TomoIdvMockStartRes"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/cn/mock/token": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvMockIssueTokenReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/TomoIdvMockIssueTokenRes"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
...
        "/v1/idv/cn/start": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvStartReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/TomoIdvStartRes"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/cn/token": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/TomoIdvIssueTokenReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/TomoIdvIssueTokenRes"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/jp/cookie/start": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/LiquidStartIdvRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/LiquidIntegratedAppResponse"
                                }
                            }
                        },
                        "description": "",
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid `body`"
                    }
                }
            }
        },
...
        "/v1/idv/jp/kyc/get": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/LiquidGetKycReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/LiquidGetResultResp"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
        "/v1/idv/jp/kyc/put": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/LiquidPutKycReq"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {}
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body`"
                    }
                }
            }
        },
        "/v1/idv/jp/notification": {
            "post": {
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/JsonValue"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/EitherStringValue"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body`"
                    }
                }
            }
        },
        "/v1/idv/jp/start": {
            "post": {
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json;charset=utf-8": {
                            "schema": {
                                "$ref": "#/components/schemas/LiquidStartIdvRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json;charset=utf-8": {
                                "schema": {
                                    "$ref": "#/components/schemas/LiquidIntegratedAppResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid `body` or `Authorization`"
                    }
                }
            }
        },
...
    }
}